Feat: CommentService 초기 구현 완료(유저와 어드민의 수정/삭제를 분리하였습니다.)

+++ src/main/java/com/sparta/spartapost/service/PostService.java
@@ -65,6 +65,7 @@ public class PostService {
 
     @Transactional
     public void adminDeletePost(Long id) {
+        if(!postRepository.existsById(id)) throw new PostNotExistException();
         postRepository.deleteById(id);
     }
 }
